refactor(slack): import get_settings from app.lib.settings
Re-route the SECRET_KEY reads in the slack token-encryption and OAuth-state
helpers (and the test cache-clear fixture) through the public settings API at
app.lib.settings instead of app.core.config. No behaviour change.
